+++ src/Himari/Prelude/Category.hs
@@ -1,0 +1,6 @@
+-- | "Control.Category" re-exports, hiding symbols that conflict with "Prelude".
+module Himari.Prelude.Category
+  ( module Export
+  ) where
+
+import Control.Category as Export hiding (id, (.))