+++ src/Himari/Prelude/Generics.hs
@@ -1,0 +1,6 @@
+-- | "GHC.Generics" re-exports, hiding symbols that conflict with "Control.Lens".
+module Himari.Prelude.Generics
+  ( module Export
+  ) where
+
+import GHC.Generics as Export hiding (from, to)